tests/wpt/tests/css/css-conditional/container-queries/size-container-no-principal-box.html

<!doctype html>
<title>CSS Container Queries Test: size container types apply to elements without a principal box</title>
<link rel="help" href="https://drafts.csswg.org/css-conditional-5/#size-container">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/cq-testcommon.js"></script>
<style>
  #outer {
    container-type: inline-size;
  }
  #inner_none {
    display: none;
    container-type: inline-size;
  }
  #inner_contents {
    display: contents;
    container-type: inline-size;
  }
  @container (min-width: 0) {
    span { color: red; }
  }
  @container (min-width: 0) {
    #ref { color: green; }
  }
  @container not (max-width: 0) {
    span { background-color: red; }
  }
  @container not (max-width: 0) {
    #ref { background-color: green; }
  }
</style>
<div id="outer">
  <div id="ref"></div>
  <div id="inner_none"><span></span></div>
  <div id="inner_contents"><span></span></div>
</div>
<script>
  setup(() => assert_implements_size_container_queries());

  test(() => {
    assert_equals(getComputedStyle(ref).color, "rgb(0, 128, 0)");
  }, "(min-width: 0) can match a container with a principal box");

  test(() => {
    assert_equals(getComputedStyle(inner_none.firstChild).color, "rgb(0, 0, 0)");
  }, "(min-width: 0) does not match a container without a principal box (display:none)");

  test(() => {
    assert_equals(getComputedStyle(inner_contents.firstChild).color, "rgb(0, 0, 0)");
  }, "(min-width: 0) does not match a container without a principal box (display:contents)");

  test(() => {
    assert_equals(getComputedStyle(ref).backgroundColor, "rgb(0, 128, 0)");
  }, "not (max-width: 0) can match a container with a principal box");

  test(() => {
    assert_equals(getComputedStyle(inner_none.firstChild).backgroundColor, "rgba(0, 0, 0, 0)");
  }, "not (max-width: 0) does not match a container without a principal box (display:none)");

  test(() => {
    assert_equals(getComputedStyle(inner_contents.firstChild).backgroundColor, "rgba(0, 0, 0, 0)");
  }, "not (max-width: 0) does not match a container without a principal box (display:contents)");
</script>
